+++ modules/client/src/multisig/internal/utils.ts
@@ -1,4 +1,8 @@
-import { hexToBytes } from "@aragon/sdk-common";
+import {
+  decodeProposalIdSubgraph,
+  encodeProposalId,
+  hexToBytes,
+} from "@aragon/sdk-common";
 import {
   DaoAction,
   ProposalMetadata,
@@ -26,8 +30,9 @@ export function toMultisigProposal(
   const endDate = new Date(
     parseInt(proposal.startDate) * 1000,
   );
+  const decodedProposalId = decodeProposalIdSubgraph(proposal.id);
   return {
-    id: proposal.id,
+    id: encodeProposalId(decodedProposalId.pluginAddress, decodedProposalId.id),
     dao: {
       address: proposal.dao.id,
       name: proposal.dao.subdomain,
@@ -66,8 +71,9 @@ export function toMultisigProposalListItem(
   proposal: SubgraphMultisigProposalListItem,
   metadata: ProposalMetadata,
 ): MultisigProposalListItem {
+  const decodedProposalId = decodeProposalIdSubgraph(proposal.id);
   return {
-    id: proposal.id,
+    id: encodeProposalId(decodedProposalId.pluginAddress, decodedProposalId.id),
     dao: {
       address: proposal.dao.id,
       name: proposal.dao.subdomain,