+++ modules/client/test/integration/tokenVoting-client/methods.test.ts
@@ -664,7 +664,7 @@ describe("Token Voting Client", () => {
         const ctxPlugin = ContextPlugin.fromContext(ctx);
         const client = new TokenVotingClient(ctxPlugin);
 
-        const proposalId = TEST_NON_EXISTING_ADDRESS + "_0x0";
+        const proposalId = TEST_NON_EXISTING_ADDRESS + "_0x0000000000000000000000000000000000000000000000000000000000000000";
         const proposal = await client.methods.getProposal(proposalId);
 
         expect(proposal === null).toBe(true);