+++ modules/common/CHANGELOG.md
@@ -19,6 +19,7 @@ TEMPLATE:
 ## [UPCOMING]
 ### Changed
 - Changes `encodeProposalId` and `decodeProposalId` function to support the new format
+- Adds `encodeProposalIdSubgraph` and `decodeProposalIdSubgraph` function to support the subgraphs proposalId format
 
 ## 0.11.0-beta
 - Adds `encodeProposalId(addr, nonce)` and `decodeProposalId(pid)`